requirements.txt
botocore
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0


pypdf==6.0.0
python-docx==1.2.0
docx2txt==0.9
olefile==0.47

langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
langchain-openai==0.3.33
langchain-text-splitters==0.3.11

httpx==0.28.1
httpcore==1.0.9

python-dotenv==1.0.1
faiss-cpu==1.11.0
sentence-transformers==5.1.0
boto3==1.40.28
numpy==1.24.3
